[Issue 79] Clean compile for gcc-4.3 with ubuntu 9 (32 bit)

+++ admb/branches/trunk-issue-78/src/linad99/ddlist.cpp
@@ -87,7 +87,7 @@ dlink * dlist::create()
                                  // zero them out
   nlinks+=1;
 
-  if (nlinks > gradient_structure::MAX_DLINKS)
+  if ((int)nlinks > gradient_structure::MAX_DLINKS)
   {
     cerr << "Need to increase the maximum number of dlinks" << endl;
 #if defined(AD_DEMO)
