+++ admb/branches/trunk-issue-78/src/linad99/df_file.cpp
@@ -93,7 +93,7 @@
       cerr << "Error trying to allocate memory for DF_FILE buffer"<<endl;
       ad_exit(1);
     }
-    for (int i=0;i<sz;i++)
+    for (unsigned int i=0;i<sz;i++)
     {
       buff[i]='\0';
     }
@@ -282,14 +282,14 @@ void DF_FILE::read_cmpdif_stack_buffer(my_off_t & lpos)
     //cout << " trying to read buff_size = " << buff_size 
       //   << " from cmpdif file" << endl;
   //cout << "offset before read is " << lseek(file_ptr,0,SEEK_CUR)<< endl;
-  if (read(file_ptr,buff,buff_size)<buff_size)
+  if (read(file_ptr,buff,buff_size)<(int)buff_size)
   {
     cerr << "End of file trying to read "<< cmpdif_file_name << endl;
     ad_exit(1);
   }
   lpos = lseek(file_ptr,-((long int) buff_size),SEEK_CUR);
   //cout << "offset after read is " << lseek(file_ptr,0,SEEK_CUR)<< endl;
-  for(int i=0;i<sizeof(unsigned int);i++)
+  for(unsigned int i=0;i<sizeof(unsigned int);i++)
   {
      fourb[i] = *(buff+buff_end+1+i);
   }
@@ -303,12 +303,12 @@ void DF_FILE::read_cmpdif_stack_buffer(my_off_t & lpos)
          //<< " into cmpdif file" << endl;
     //cout << "offset before write is " << lseek(file_ptr,0,SEEK_CUR)<< endl;
     //if (write(file_ptr,buff,buff_size)<buff_size)
-    for(int i=0;i<sizeof(unsigned int);i++)
+    for(unsigned int i=0;i<sizeof(unsigned int);i++)
     {
 	 *(buff+buff_end+1+i)=fourb[i]; // save the offset at the
 			//end of the used part of the buffer
     }
-    if (write(file_ptr,buff,buff_size)<buff_size)
+    if (write(file_ptr,buff,buff_size)<(int)buff_size)
     {
       cerr << "End of file trying to write to file "<< cmpdif_file_name << endl;
       cerr << "There is probably no more room on the TMP1 (if defined) device\n"
